LOG-001 Split up inner classes to their own class.
